requirements.txt
pystray
pillow
requests
fastmcp
psutil
uvicorn
fastapi
